refactor(ui): AdaptiveDropdown is the suite's, not mail's
The calendar's phone form asks what this component was built to answer: a row
whose value is one of a few, which is a dropdown on a desktop and has to be a
sheet on a phone. It moves to src/components rather than being reached for
across apps or, worse, built a second time — a sheet made to look like this one
is a sheet that has to keep looking like it.

stripShortcutHint travels with it, to src/utils/actionLabel: the component reads
labels through it, and there is nothing about mail in a regex that trims "(E)"
off the end of one. useScreenSize now comes from the shared composable it was
already re-exported from.
