+++ frontend/src/apps/mail/utils/index.ts
@@ -13,12 +13,6 @@ import VideoIcon from '@/apps/mail/components/Icons/VideoIcon.vue'
 
 import type { ComposeMailData, MailboxData, Recipient } from '@/apps/mail/types'
 
-// Keyboard hints in action labels — "Archive Thread (E)", "Move to Trash (Delete)" —
-// are noise on touch surfaces. Strips only trailing parentheticals that look like
-// shortcuts, so a folder named "Work (old)" is never clipped.
-const SHORTCUT_HINT =
-	/\s*\((?:(?:Shift|Ctrl|Cmd|Alt|⌘|⇧|⌥)\+)*(?:[A-Z!,.;]|Delete|Backspace|Esc(?:ape)?|Enter|Tab|Space|↑\/K|↓\/J)\)$/
-export const stripShortcutHint = (label: string) => label.replace(SHORTCUT_HINT, '')
 
 export const toTitleCase = (str: string) =>
 	str
